Normalize line endings
We want to achieve a higher reproducibility standard on Windows: if a
Windows user checks out this repository and compiles a `rel/x.x.x` tag,
he should obtain exactly the same JAR files as those published.

Therefore we need to normalize the line endings of all text resources to
LF. This closes #29.