Remove `1.x/xsd/changelog.xsd`
This was established to publish XSDs along with releases. Though,
this turned out to be not a good idea: versioning of XSDs and
releases are not correlated. That is, imagine you make a breaking
change to `log4j-changelog`, though XSDs are untouched. Due to the
breaking change, you need to bump the major version. As a result,
the XSD will be moved to `2.x/xsd/changelog.xsd`. Though this is
does not make sense, since XSDs have not changed.